Secured the beer views

+++ beers/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+
 from django.db.models import Q
 
 from .models import Beer, Category
@@ -51,7 +53,12 @@ def beer_detail(request, beer_id):
     return render(request, 'beers/beer_detail.html', context)
 
 
+@login_required
 def add_beer(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only 12 Acres owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = BeerForm(request.POST, request.FILES)
         if form.is_valid():
@@ -71,8 +78,13 @@ def add_beer(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_beer(request, beer_id):
     """ Edit a beer in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only 12 Acres owners can do that.')
+        return redirect(reverse('home'))
+
     beer = get_object_or_404(Beer, pk=beer_id)
     if request.method == 'POST':
         form = BeerForm(request.POST, request.FILES, instance=beer)
@@ -95,8 +107,13 @@ def edit_beer(request, beer_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_beer(request, beer_id):
     """ Delete a beer """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only 12 Acres owners can do that.')
+        return redirect(reverse('home'))
+
     beer = get_object_or_404(Beer, pk=beer_id)
     beer.delete()
     messages.success(request, 'Beer deleted!')
